+++ lib/src/readers/id3v2.dart
@@ -187,7 +187,7 @@ class ID3V2Reader extends Reader {
   // [UTF-8] encoded Unicode [UNICODE]. Terminated with $00.
   static const _utf8 = 0x03;
 
-  static const _headerLength = 10;
+  int get _headerLength => version_o2 > 2 ? 10 : 6;
 
   int version_o1 = 2;
   int version_o2 = 0;
@@ -211,9 +211,14 @@ class ID3V2Reader extends Reader {
     version_o2 = sBytes[3];
     version_o3 = sBytes[4];
 
-    // use in future
+    final flags = sBytes[5];
+
     // ignore: unused_local_variable
-    final flag = sBytes[5];
+    final unsync = flags & 0x80 != 0;
+    // ignore: unused_local_variable
+    final xheader = flags & 0x40 != 0;
+    // ignore: unused_local_variable
+    final experimental = flags & 0x20 != 0;
 
     final size = _sizeOf(sBytes.sublist(6, 10));
 
@@ -222,8 +227,8 @@ class ID3V2Reader extends Reader {
     var end = true;
 
     while (end) {
-      final tag = utf8.decode(sBytes.sublist(offset, offset + 4));
       final encoding = _getEncoding(sBytes[offset + _headerLength]);
+      final tag = encoding.decode(sBytes.sublist(offset, offset + 4));
 
       final len = _sizeOf(sBytes.sublist(offset + 4, offset + 8));
 
@@ -236,11 +241,8 @@ class ID3V2Reader extends Reader {
                 encoding);
             break;
           default:
-            tags[_getTag(tag)] = encoding.decode(sBytes
-                .sublist(
-                    offset + _headerLength + 1, offset + _headerLength + len)
-                .where((i) => i != 0)
-                .toList());
+            tags[_getTag(tag)] = encoding.decode(_clearFrameData(sBytes.sublist(
+                offset + _headerLength + 1, offset + _headerLength + len)));
         }
       }
 
@@ -249,6 +251,13 @@ class ID3V2Reader extends Reader {
     }
 
     return tags;
+  }
+
+  List<int> _clearFrameData(List<int> bytes) {
+    if (bytes.length > 3 && bytes[0] == 0xFF && bytes[1] == 0xFE) {
+      bytes = bytes.sublist(2);
+    }
+    return bytes.where((i) => i != 0).toList();
   }
 
   int _sizeOf(List<int> block) {
@@ -328,7 +337,7 @@ class UTF16 extends Encoding {
 class _UTF16Decoder extends Converter<List<int>, String> {
   @override
   String convert(List<int> input) {
-    return String.fromCharCodes(input);
+    return new String.fromCharCodes(input);
   }
 }